% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Tools.R
\name{bc_drive}
\alias{bc_drive}
\title{Sets a general file location so different users don't have to change their R Scripts. Note that this works with ProjectWise Drive
and OneDrive when you have connected projects to your local system.}
\usage{
bc_drive(ProjectFolder, Drive = "OneDrive")
}
\arguments{
\item{ProjectFolder}{Specify the rest of the file path in quotes}

\item{Drive}{Drive type, choose between "OneDrive", "ProjectWise", and "PersonalOneDrive"}
}
\description{
Sets a general file location so different users don't have to change their R Scripts. Note that this works with ProjectWise Drive
and OneDrive when you have connected projects to your local system.
}
\examples{
bc_drive("Reuse Pilot/Data", "Personal One Drive")
bc_drive("000000 - Project Name/001 Task Name/01 Subfolder", "projectwise")

}
