% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bctools.R
\name{bcverbose}
\alias{bcverbose}
\title{Set default verbosity level for bytecode \emph{print} method}
\usage{
bcverbose(lvl = NULL)
}
\arguments{
\item{lvl}{verbosity level ( 0 or 1 or 2) - optional
if setted - set default bytecode verbosity level
  0 - display only source references ( if they are available, if they aren't print expression references instead )
  1 - the same as 0 + display bytecode version and display expression references ( if they are available )
  2 - the same as 1 + display every operand's argument ( including ones used just for debugging )}
}
\value{
current verbosity level
}
\description{
\code{bcverbose} Set default verbosity level for bytecode \emph{print} method
}
\examples{

library(compiler)
library(bctools)
a <- function(x){
  r <- 1
  while(x){
    r = r + x*x
    x = x-1
  }
  r
}
bc <- compiler::cmpfun(a)

#set default verbosity level
bcverbose(2);
disassemble(bc);

}
