% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{export_plot}
\alias{export_plot}
\title{Creates a png file of plots with automatic scaling for adding to word documents}
\usage{
export_plot(plot, filename = "DRAFT R Plot", vscale = 1, hwidth = 6.5)
}
\arguments{
\item{plot}{Plot to be exported}

\item{filename}{Desired file name in quotes without file extension.}

\item{vscale}{Default height is 4", based on typical plot ratios to copy into a word doc.  Increase scale to multiply size.}

\item{hwidth}{Default width is 6.5", only change if plot is being added to a different page width. Specify in inches.}
}
\description{
Creates a png file of plots with automatic scaling for adding to word documents
}
