% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-boxandwhisker.R
\name{geom_boxandwhisker}
\alias{geom_boxandwhisker}
\title{A box and whisker quantile plot}
\usage{
geom_boxandwhisker(
  mapping = NULL,
  data = NULL,
  stat = "boxandwhisker",
  position = "dodge2",
  quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
  count = TRUE,
  middlepoint = "mean",
  whiskerbar = FALSE,
  outliers = TRUE,
  ...,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = NULL,
  outlier.size = NULL,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  whisker.colour = NULL,
  whisker.color = NULL,
  whisker.linetype = NULL,
  whisker.linewidth = NULL,
  staple.colour = NULL,
  staple.color = NULL,
  staple.linetype = NULL,
  staple.linewidth = NULL,
  median.colour = NULL,
  median.color = NULL,
  median.linetype = NULL,
  median.linewidth = NULL,
  box.colour = NULL,
  box.color = NULL,
  box.linetype = NULL,
  box.linewidth = NULL,
  count.colour = NULL,
  count.color = NULL,
  count.size = NULL,
  count.label = FALSE,
  midpoint.colour = NULL,
  midpoint.color = NULL,
  midpoint.fill = NULL,
  midpoint.shape = NULL,
  midpoint.size = NULL,
  midpoint.stroke = NULL,
  midpoint.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  staplewidth = 0.9,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{count}{Whether to display (\code{TRUE}) or hide (\code{FALSE}) the count below the plot.}

\item{middlepoint}{Can be equal to "mean" (default) or "90th" to select location, can be set to FALSE to remove.}

\item{outliers}{Whether to display (\code{TRUE}) or discard (\code{FALSE}) outliers
from the plot. Hiding or discarding outliers can be useful when, for
example, raw data points need to be displayed on top of the boxplot.
By discarding outliers, the axis limits will adapt to the box and whiskers
only, not the full data range. If outliers need to be hidden and the axes
needs to show the full data range, please use \code{outlier.shape = NA} instead.}

\item{outlier.colour, outlier.color, outlier.fill, outlier.shape, outlier.size, outlier.stroke, outlier.alpha}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{whisker.colour, whisker.color, whisker.linetype, whisker.linewidth}{Default aesthetics for the whiskers. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{staple.colour, staple.color, staple.linetype, staple.linewidth}{Default aesthetics for the staples. Set to \code{NULL} to inherit from the
data's aesthetics. Note that staples don't appear unless the \code{staplewidth}
argument is set to a non-zero size.}

\item{median.colour, median.color, median.linetype, median.linewidth}{Default aesthetics for the median line. Set to \code{NULL} to inherit from the
data's aesthetics. **** Not yet implemented, waiting on next ggplot2 release.}

\item{box.colour, box.color, box.linetype, box.linewidth}{Default aesthetics for the boxes. Set to \code{NULL} to inherit from the
data's aesthetics.**** Not yet implemented, waiting on next ggplot2 release.}

\item{count.colour, count.color, count.size}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{count.label}{Whether to display the string "n=" before the count. Defaults to \code{FALSE}.}

\item{midpoint.colour, midpoint.color, midpoint.fill, midpoint.shape, midpoint.size, midpoint.stroke, midpoint.alpha}{Default aesthetics for middlepoint. Set to \code{NULL} to use plot defaults.}

\item{notch}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{notchwidth}{For a notched box plot, width of the notch relative to
the body (defaults to \code{notchwidth = 0.5}).}

\item{staplewidth}{The relative width of staples to the width of the box.Defaults to 0.9.
Staples mark the ends of the whiskers with a line.}

\item{varwidth}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}

\item{geom, stat}{Use to override the default connection between
\code{geom_boxplot()} and \code{stat_boxplot()}. For more information about
overriding these connections, see how the \link[=layer_stats]{stat} and
\link[=layer_geoms]{geom} arguments work.}

\item{qs}{Can be used to change the quantiles displayed in the plot. Defaults to \code{c(.05,.25,.50,.75,.95)}}
}
\description{
The boxplot compactly displays the distribution of a continuous variable.
It visualises five summary statistics (the median, two hinges
and two whiskers), and all "outlying" points individually.
}
\section{Summary statistics}{

By default, the box is the same as \code{ggplot2::geom_boxplot}, but the whiskers are based on quantiles instead of IQR.
Calculations use the 5 quantiles provided, with whiskers extending to the first and last quantiles, the hinges
corresponding to the second and fourth, and middle line for the third.
By default, it also includes a middle point at the mean and displays the count below the plot.

In a notched box plot, the notches extend \code{1.58 * IQR / sqrt(n)}.
This gives a roughly 95\% confidence interval for comparing medians.
See McGill et al. (1978) for more details.
}

