% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{geom_boxplot_q}
\alias{geom_boxplot_q}
\title{A box and whiskers quantile plot}
\usage{
geom_boxplot_q(
  mapping = NULL,
  data = NULL,
  stat = "boxplotq",
  position = "dodge2",
  quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
  count = TRUE,
  middlepoint = "mean",
  whiskerbar = FALSE,
  ...,
  outliers = TRUE,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  staplewidth = 0.9,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{outliers}{Whether to display (\code{TRUE}) or discard (\code{FALSE}) outliers
from the plot. Hiding or discarding outliers can be useful when, for
example, raw data points need to be displayed on top of the boxplot.
By discarding outliers, the axis limits will adapt to the box and whiskers
only, not the full data range. If outliers need to be hidden and the axes
needs to show the full data range, please use \code{outlier.shape = NA} instead.}

\item{outlier.colour, outlier.color, outlier.fill, outlier.shape, outlier.size, outlier.stroke, outlier.alpha}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{notch}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{notchwidth}{For a notched box plot, width of the notch relative to
the body (defaults to \code{notchwidth = 0.5}).}

\item{staplewidth}{The relative width of staples to the width of the box.
Staples mark the ends of the whiskers with a line.}

\item{varwidth}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}

\item{geom, stat}{Use to override the default connection between
\code{geom_boxplot()} and \code{stat_boxplot()}. For more information about
overriding these connections, see how the \link[=layer_stats]{stat} and
\link[=layer_geoms]{geom} arguments work.}

\item{whisker.colour, whisker.color, whisker.linetype, whisker.linewidth}{Default aesthetics for the whiskers. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{median.colour, median.color, median.linetype, median.linewidth}{Default aesthetics for the median line. Set to \code{NULL} to inherit from the
data's aesthetics.}

\item{staple.colour, staple.color, staple.linetype, staple.linewidth}{Default aesthetics for the staples. Set to \code{NULL} to inherit from the
data's aesthetics. Note that staples don't appear unless the \code{staplewidth}
argument is set to a non-zero size.}

\item{box.colour, box.color, box.linetype, box.linewidth}{Default aesthetics for the boxes. Set to \code{NULL} to inherit from the
data's aesthetics.}
}
\description{
The boxplot compactly displays the distribution of a continuous variable.
It visualises five summary statistics (the median, two hinges
and two whiskers), and all "outlying" points individually.
}
\section{Summary statistics}{

The lower and upper hinges correspond to the first and third quartiles
(the 25th and 75th percentiles). Whiskers extend to the specified quantiles, unlike \code{ggplot2::geom_boxplot}, which
calculates outliers.

In a notched box plot, the notches extend \code{1.58 * IQR / sqrt(n)}.
This gives a roughly 95\% confidence interval for comparing medians.
See McGill et al. (1978) for more details.
}

