% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{geom_corrosion_heatmap}
\alias{geom_corrosion_heatmap}
\title{Lead contour plot}
\usage{
geom_corrosion_heatmap(
  ph_range = seq(6, 10, 0.1),
  alk_range = seq(1, 50, 1),
  tds_average = 50,
  ...
)
}
\arguments{
\item{ph_range}{Set the pH range, defaults to seq(6, 10, 0.1).}

\item{alk_range}{Set the pH range, defaults to seq(1, 50, 1).}

\item{tds_average}{Average tds of your data, defaults to 50 mg/L.}
}
\description{
Generates a heat map of lead solids and a transition line between controlling lead solids.
}
\examples{
library(tidywater)
ggplot() +
  geom_corrosion_heatmap() +
  geom_text(x =10, y = 8, label = "Hydrocerussite", color = "white", size = 5) +
  geom_text(x = 10, y = 6.5, label = "Cerussite", color = "white", size = 5) +
  theme_bc()+
  labs(fill = "log Pb Conc\n(mg/L)", x = "Alkalinity (mg/L)", color = "", y = "pH", shape = "") +
  coord_cartesian(xlim = c(6, 40),ylim = c(6,10))+
  theme(legend.position = "bottom")

ggplot() +
  geom_corrosion_heatmap(alk_range = seq(40,250,10))+
  geom_point(data = tidywater::water_df, aes(x = alk, y = ph, shape = "Historical Data"), color = "grey") +
  geom_text(aes(x = 100, y = 9), label = "Hydrocerussite", color = "white", size = 5) +
  geom_text(aes(x = 100, y = 7), label = "Cerussite", color = "white", size = 5) +
  theme_bc()+
  labs(fill = "log Pb Conc\n(mg/L)", x = "Alkalinity (mg/L)", color = "", y = "pH", shape = "") +
  coord_cartesian(xlim = c(40, 250),ylim = c(6,10))+
  theme(legend.position = "bottom")

}
