% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{geom_corrosion_ranges}
\alias{geom_corrosion_ranges}
\title{Automatically add typical corrosion ranges to plots
Ranges are based on this corrosion write up, already QC'd by Damon Roth (access to doc for BC employees only)
https://brwncald-my.sharepoint.com/:w:/p/lmckenna/Ebhc5hMxdZpHq5aKyoaGijQBheBR0EYLVtEEMJFaXpnARw?e=kRv8dX}
\usage{
geom_corrosion_ranges(
  data,
  alpha = 0.3,
  fill = "cyan",
  xmin = -Inf,
  xmax = Inf,
  index_column,
  ...
)
}
\arguments{
\item{data}{Data frame containing a column of named corrosion/scaling indices}

\item{alpha}{Set the opacity of the rectangles, default is 0.3}

\item{fill}{Set the color of the rectangles, default is "cyan"}

\item{index_column}{Name of the column in the plot data frame with corrosion/scaling indices names (should be the column used for facetting)}
}
\description{
Automatically add typical corrosion ranges to plots
Ranges are based on this corrosion write up, already QC'd by Damon Roth (access to doc for BC employees only)
https://brwncald-my.sharepoint.com/:w:/p/lmckenna/Ebhc5hMxdZpHq5aKyoaGijQBheBR0EYLVtEEMJFaXpnARw?e=kRv8dX
}
\examples{
corr_plot <- water_df \%>\%
  define_water_chain() \%>\%
  calculate_corrosion_once() \%>\%
  pivot_longer(aggressive:csmr, names_to = "index", values_to = "result")

ggplot(corr_plot, aes(x = ph, y = result)) +
  geom_point() +
  facet_wrap(~index) +
  geom_corrosion_ranges(data = corr_plot, index_column = "index")

}
