% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{scale_fill_bc}
\alias{scale_fill_bc}
\title{BC Color Palette with scale_fill}
\usage{
scale_fill_bc(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of color palette to use. Current options: "primary", "rainbow"}

\item{discrete}{Set equal to false if gradient is desired}

\item{reverse}{Set equal to TRUE if you want the colors in the opposite order}
}
\description{
BC Color Palette with scale_fill
}
