% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{theme_bc}
\alias{theme_bc}
\title{Theme function for BC defaults
Automatically sets BC color palette.}
\usage{
theme_bc(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{Can be set to adjust default font sizes}

\item{base_family}{Can be set to change default font (R isn't great at fonts, so be careful)}
}
\description{
Theme function for BC defaults
Automatically sets BC color palette.
}
