% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Tools.R
\name{bc_drive}
\alias{bc_drive}
\title{Sets a general file location so different users don't have to change their R Scripts. Note that this works with ProjectWise Drive
and OneDrive when you have connected projects to your local system.}
\usage{
bc_drive(projectfolder, drive = "CodeReview")
}
\arguments{
\item{projectfolder}{Specify the rest of the file path in quotes}

\item{drive}{Drive type. Defaults to the shared Code Review folder. Otherwise, choose between the following:
\itemize{
\item "CodeReview" (the default): The internally shared OneDrive folder that should typically house data files for R Scripts.
This folder should be used whenever possible to make it easier for others to run and review your code.
File path - "C:/Users/yourname/Brown and Caldwell/Sierra Johnson - GitHub Code Review Files/" or
"C:/Users/yourname/OneDrive - Brown and Caldwell/GitHub Code Review Files/" (path depends on whether you used a shortcut or sync to add this folder to your OneDrive)
\item "OneDrive" : Other folders in the file path "C:/Users/yourname/Brown and Caldwell" or "C:/Users/yourname/OneDrive - Brown and Caldwell" if you haven't synced any shared folders.
This method is recommended if you have project files only shared with certain people. (It also works with folders you've shared with others from "Shared Internally" folder).
\item "ProjectWise" : ProjectWise Drive files. Note that these are often slower than OneDrive files and are not recommended.
File path:  C:/Users/yourname/ProjectWise/Brown & Caldwell
\item "MyOneDrive" : Files in your personal OneDrive documents.
File path: C:/Users/yourname/OneDrive - Brown and Caldwell/Documents/
\item "Egnyte" : Folders in the Egnyte desktop app. App must be installed on computer for the function to work.
File path: Y:/Shared/Clients
}}
}
\description{
Sets a general file location so different users don't have to change their R Scripts. Note that this works with ProjectWise Drive
and OneDrive when you have connected projects to your local system.
}
\examples{
bc_drive("Reuse Pilot/Data", "Personal One Drive")
bc_drive("000000 - Project Name/001 Task Name/01 Subfolder", "projectwise")
bc_drive("Golden_City of-CO/153349 - Golden WTP Facility Plan/", drive = "Egnyte")

}
