% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Tools.R
\name{factor_wq}
\alias{factor_wq}
\title{Factor water quality data to produce a standard table}
\usage{
factor_wq(data, parameter_column, ...)
}
\arguments{
\item{data}{Data frame containing the column of water quality data to be factored}

\item{parameter_column}{Name of column with water quality parameter names.}
}
\description{
Factor water quality data to produce a standard table
}
\examples{
water <- water_df \%>\%
  pivot_longer(everything(), names_to = "param", values_to = "result") \%>\%
  factor_wq(param)


}
