% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tools.R
\name{geom_boxandwhisker}
\alias{geom_boxandwhisker}
\title{Box and Whisker Plot function that uses 25th and 75th percentile for the box and 5th and 95th percentiles for whiskers.
The middle point can be at the mean or 90th percentile.
Outlier points can be included or excluded.  To set fill to a single color, include the argument fill =.}
\usage{
geom_boxandwhisker(
  outlier = TRUE,
  count = TRUE,
  middlepoint = "mean",
  whiskerbar = TRUE,
  alpha = 0.8,
  width = 0.9,
  fontsize = 9,
  whiskerloc = 0.05,
  countlabel = FALSE,
  whiskerlabel = FALSE,
  boxedgelabel = FALSE,
  medianlabel = FALSE,
  ...
)
}
\arguments{
\item{outlier}{Can be set to FALSE to remove points <5th percentile and >95th percentile.}

\item{count}{Can be set to FALSE to remove count from below the whisker.}

\item{middlepoint}{Can be equal to "mean" (default) or "90th" to select location, can be set to FALSE to remove.}

\item{whiskerbar}{Can be set to FALSE to remove horizontal bars on the ends of the whiskers.}

\item{alpha}{Can be set to a different transparency or NA if an alpha scale is needed (not recommended).}

\item{width}{Can be set to a number between 0 and 1, with lower numbers increasing space between boxes.}

\item{fontsize}{Can be used to adjust the size of the count. Use a number equivalent to a standard size in points.}

\item{whiskerloc}{Can be used to adjust the percentile that the whiskers extend to, use the low value, high will be calculated.}

\item{countlabel}{Can be set to TRUE if you want the count to appear as n=#}
}
\description{
Box and Whisker Plot function that uses 25th and 75th percentile for the box and 5th and 95th percentiles for whiskers.
The middle point can be at the mean or 90th percentile.
Outlier points can be included or excluded.  To set fill to a single color, include the argument fill =.
}
