% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bctools.R
\name{print.disassembly}
\alias{print.disassembly}
\title{Print bytecode object to output and returns it \emph{invisibly} (via \code{\link{invisible}(x)})}
\usage{
\method{print}{disassembly}(x, prefix = "", verbose = 0, maxdepth = 3,
  depth = 0, ...)
}
\arguments{
\item{x}{Bytecode object to be printed}

\item{prefix}{number of spaces to print before each line ( used for intendation )}

\item{verbose}{verbosity level ( 0 or 1 or 2)}

\item{maxdepth}{Maximum depth of nested functions which are printed}

\item{depth}{Current depth of nested functions which are being printed ( used for internal purposes in print recursion )}

\item{...}{Numeric, complex, or logical vectors.}
}
\description{
\code{print.disassembly} print bytecode object into output in human-friendly way.
}
\details{
This is implementation of print method for bytecode object. 
It works under internal R Bytecode structure.
}
\examples{
library(compiler)
library(bctools)
a <- function(x){
  r <- 1
  while(x){
    r = r + x*x
    x = x-1
  }
  r
}
bc <- compiler::cmpfun(a)

print(compiler::disassemble(bc));
print(compiler::disassemble(bc), verbose=1);
print(compiler::disassemble(bc), verbose=2);

}
